Time: 2573 ms (5.01%), Space: 545.8 MB (5.04%) - LeetHub

+++ 3510-minimum-pair-removal-to-sort-array-ii/3510-minimum-pair-removal-to-sort-array-ii.cpp
@@ -27,7 +27,7 @@ public:
                 mp[i]=true;
         }
         node* cur = new node();
-        cur->idx = nums.size()-1;
+        cur->idx = (int)nums.size()-1;
         cur->prev = prev;
         cur->val = nums[nums.size()-1];
         if(prev)
